Added a new OrganisationProposal model.
PT-1071 PT-1083. When a member of a 3rd party organisation registers to the API from the providers UI, he can make a proposal to add a new 3rd party organisation.Since the process is still quite unclear, the proposals and and 3rd party organisation wishes can be collected with this model and stored in database as detached. The use case is that an admin can see which organisation the new user likes to represent and then the real organisation.Organisation instance can be created (if necessary and not yet done) and linked to the user.
Added tests for the OrganisationProposal model.

PT-1074 PT-1083.