Fix to events list meta count
LinkedEvents paginates the results, but Kultus API filters the paginated sets. This leads to a situation where the events count in meta data, easily does not match with the fact. If events.meta is used, the count is bigger than the filtered result. If a length of  integrated events is used, it would work only for unpaginated results. Using meta given by LinkedEvents is better, but can lead to empty pages!

PT-1170