Occurrence signal and event republish checks whether it should update or not.
Fetches data from LinkedEvents API, because the event start time and end time are not stored anywhere else.

The changes of the time fields should be checked, to prevent duplicate signal calls (1 from create, 1 from relation updates), but in a post_save-signal the information is not available and in a pre_save-signal the occurrences are not yet persisted, so it would be a pain to republish by using send_event_republish (and p_event.get_end_time_from_occurrences())

Some kind of effective cache should be used.

PT-1200 PT-1227.