Refactored Occurrence signals to an overridden save and delete in model.
When using an overridden save and delete function instead of post_save signal, also the old object value are available and can be compared when resolving whether or not the event should be updated. Signals got called twice: once when created and second time when the relations updated - Overridden model function works better.

PT-1200 PT-1227.