Refactored send_enrolment_summary_report_to_providers
Enrolment used EnrolmentQuerySet as a manager, which lead to a situation where notification sending logic was part of the query set method. A query set should not have a business logic inside it, but only return a query set, so created a new EnrolmentManager to manage the notification sending. However, its not even managers job to maintain the logic for notification sending, so moved the sending logic to a new service module.

PT-1364.