Implement support for Person deletion
Implemented changes that allow Person objects to be deleted even when
they are related to an Enrolment or a StudyGroup. The Enrolment and the
StudyGroup won't be affected by the deletion. The need for this comes
from GDPR where users can request their personal data to be deleted, and
too old personal data should be deleted automatically.