Implement support for event contact info deletion
Implemented a method for deleting a `PalvelutarjotinEvent`'s contact
info, so in practice fields `contact_person`, `contact_email` and
`contact_phone_number`. There is also a new datetime field
`contact_info_deleted_at` which is populated when the deletion takes
place. Note: the deletion method needs to be executed explicitly, just
emptying the fields will not be considered as a contact info deletion in
this context.

The event contact info deletion process is executed automatically when
the contact person `Person` object, or the `User` object related to
that, is deleted.