Fail safer LinkedEvents-events hydration to the enrolment reports
PT-1525 PT-1527.
Instead of throwing errors, just log them with a logger,
so that more data is fetched. Instead of giving up
when a first error occurs, we should try to fetch
the rest of the info that is given.
Also, an improvement to the logging was added,
so that it's clearer where the issue is.
It should also be noted that not all the places have coordinates,
e.g. `helsinki:internet` for internet events,
so we should not throw an KeyError in these cases,
because it's a valid case.