Add EventQueueEnrolment model and an EnrolmentBase class
PT-1651.

The EnrolmentBase class holds the common fields that
are used in both, when enrolling to an occurrence or
when enrolling to the event enrolment queue.

The idea with the new event queue model is, 
that an user can enrol to the event queue 
when the occurrences that he really 
wanted to enrol, are full. Then, if any spots are left free
later on, an admin can contact the queue contacts
and take the queue information and 
make an occurrence enrolment out of those.