Add EnrolEventQueueMutation to handle event queue enrolments over API
PT-1651.

Added a new input base for the event queue enrolments and the
real (old) occurrence enrolments, so that they would share most
of the inputs. Refactored the `EnrolOccurrenceMutation` to use same
Input base as the new `EnrolEventQueueMutation` uses.

Add a new Enum for different queue statuses.

Rename EventQueueEnrolment status getter and annotate it as property.