Add an occurrence filter to enrolments query
PT-1651.

There was an enrolment query, but it did not have any tests.
Now the enrolment query has the test coverage and also a new
occurrence filter, because it does not make much sense to just get
a list of all the enrolments. The enrolments can now also be sorted.