feat: event queue enrolment counts are included in the enrolment summary
PT-1717.

The enrolment summary now contains
1. a total new event queue enrolment count
2. event specific new queue enrolment count
Both of the new infos are only from events that are related to the recipient.

The `send_enrolment_summary_report_to_providers` is now removed
from the Enrolment model as a needless function.
It is replaced with an util that handles the both: the event specific
event queue enrolments and the event occurrence specific enrolments.