fix: remove port from the response of `get_client_ip`
PT-1909.

> The error message "invalid input syntax for type inet:
"83.148.233.59:39798"" indicates that you are trying to store or use the
string 83.148.233.59:39798 in a database column or a program variable
that is expecting a value of the inet data type. The inet data type is
typically used to store IPv4 or IPv6 addresses, optionally including a
subnet mask. It does not include port numbers. The problem is the
presence of the port number :39798 at the end of the IP address.

In latest version of django-auditlog, v.3.1.1, they have splitted the
remote address and port in 2 different fields. This is not the case yet
in the version v.3.0.0 that we are using. Since the field is a Django's
GenericIPAddressField, the port should not be included in the field
value.

The `get_client_ip` utility function did not remove the port from the
remote address, but it can sometimes be included in the
`request.META.REMOTE_ADDR`, so it needs to be explicitly removed. It is
also removed by the django-auditlog middleware, that is using a similar
kind of (private) function for similar purpose.